+++ rbx_xml/CHANGELOG.md
@@ -1,5 +1,10 @@
 # rbx_xml Changelog
 
+## Unreleased
+* Added support for `Terrain.SmoothGrid`, and adjacent support for `TerrainRegion.SmoothGrid`. ([#444])
+
+[#444]: https://github.com/rojo-rbx/rbx-dom/pull/444
+
 ## 2.0.0 (2025-10-10)
 * Upgrade rbx-dom dependencies, which results in breaking changes to some data types.
 * Implement support for serializing and deserializing the `NetAssetRef` type. ([#555])